Put findpts in a seperate directory from ogs
The only modifications to existing files are now a few added paths in
CMake files
