+++ 3rd_party/gslib/ogs/include/ogsKernels.hpp
@@ -184,9 +184,6 @@ namespace ogs {
   extern occa::kernel scatterManyKernel_int;
   extern occa::kernel scatterManyKernel_long;
 
-  extern occa::kernel findpts_local_eval_2;
-  extern occa::kernel findpts_local_eval_3;
-
   extern occa::stream defaultStream;
   extern occa::stream dataStream;
 
@@ -195,8 +192,6 @@ namespace ogs {
   extern occa::properties kernelInfo;
 
   void freeKernels();
-
-  occa::kernel initFindptsKernel(MPI_Comm comm, occa::device device, dlong D, const dlong *n);
 }
 
 void occaGatherScatter(const  dlong Ngather,
