+++ src/ogsFindpts/ogsDevFindpts.c
