+++ src/ogsFindpts/ogsHostFindpts.c
