+++ src/ogsFindpts/ogs_findpts.c
