+++ src/ogsFindpts/ogs_findpts_local.cpp
@@ -1,7 +1,7 @@
 
 #include <type_traits>
 #include "ogstypes.h"
-#include "ogs.hpp"
+#include "ogs_FINDPTS.hpp"
 
 extern "C" {
