+++ src/udf/CMakeLists.txt
@@ -35,6 +35,7 @@ set(INCLUDE_DIRS
   ${NEKRS_INSTALL_DIR}/include/cds
   ${NEKRS_INSTALL_DIR}/include/linAlg
   ${NEKRS_INSTALL_DIR}/include/interp
+  ${NEKRS_INSTALL_DIR}/include/ogsFindpts
   ${NEKRS_INSTALL_DIR}/occa/include
 )
